refactor(mcp): address review feedback on scope enforcement
- Share the audit-mutation lifecycle: extract runMcpMutation so survey and
  workflow runners no longer duplicate the try/catch/status/queue logic.
- registerScopedTool now takes a non-empty scope tuple ([string, ...string[]])
  so a tool cannot be registered gating on an empty scope set.
- Route-level workflow scope-denial test asserts no audit log is built or queued
  (proves the gate fires before any side effect).
- Narrow the insufficient-scope assertion type in the guard-scopes unit test.
- Document that test_workflow needs workspace write/manage despite its read scope.
