+++ docs/platform/mcp/setup.mdx
@@ -154,8 +154,9 @@ periodically for safety.
   <Accordion title="A tool call fails even though you're connected">
     Scopes don't override your workspace role. Creating, editing, or deleting surveys or workflows
     requires `write` or `manage` permission on that workspace — ask a workspace owner/manager to grant
-    it. A `:write` tool also 403s if your token was only granted the matching `:read` scope; reconnect
-    and approve write access.
+    it. `test_workflow` also needs `write`/`manage` even though it's a read-scoped dry-run, because it
+    evaluates the live workflow definition. A `:write` tool likewise 403s if your token was only granted
+    the matching `:read` scope; reconnect and approve write access.
   </Accordion>
   <Accordion title="Self-hosted: the browser can't complete sign-in / redirects fail">
     OAuth discovery and redirects are built from your configured base URL. Serve Formbricks from a
